requirements.txt
apache-beam
apache_beam[gcp]
google-api-python-client
google-auth
apache-airflow-providers-apache-beam
apache-airflow-providers-google